code fix not ready for production.
git-svn-id: https://fluentdata.svn.codeplex.com/svn@22233 aadf2aa7-86a3-4242-a263-039aa7a476c3

+++ Source/Main/FluentData.IntegrationTests/Providers/Sqlite/SqliteTests.cs
@@ -98,7 +98,7 @@ namespace FluentData.Providers.Sqlite
 
 		public void Custom_mapper_using_dynamic(Product product, dynamic row)
 		{
-			product.ProductId = row.ProductId;
+			product.ProductId = (int) row.ProductId;
 			product.Name = row.Name;
 		}
